sutrofm/serializers.py
from sutrofm.models import User
from rest_framework import serializers


class UserSerializer(serializers.HyperlinkedModelSerializer):
  class Meta:
    model = User
    fields = ['username', 'parties']
