requirements.txt
redis==3.4.1
asgiref==3.2.7
Django==3.0.4
psycopg2==2.8.4
pytz==2019.3
sqlparse==0.3.1
social-auth-app-django==3.1.0
simplejson==3.17.0
python-dateutil==2.8.1
psutil==5.7.0
django-websocket-redis==0.6.0
spotipy==2.10.0
djangorestframework==3.11.0
django-model-utils==4.0.0
